Both Zybo-Z7 and Artix-7 projects now build
The full design is untested in hardware, but the HDMI input logic is
known to work on the Zybo-Z7.

A generic was added to control clock buffering.  These projects require
using a BUFG (since HDMI input and output are in different I/O banks),
but better performance can be had by using BUFIO/BUFR.

Signed-off-by: Charles Steinkuehler <cstein@newtek.com>